Move the event struct to eventloop module
